Use ELO deltas average as a final ELO delta, and increase K
The old way had too much volatility, meaning a player can gain too much
ELO points in 5 minute (sometime 60 or more), or loose to much.  To
solve that, instead of using the sum of ELO deltas as a final ELO delta,
it now does the average of ELO deltas.

Betwen each update, en ELO score can now change +- 10 points (because
the "K" factor is 10).  It's not a lot so change it to 25, the value
used in chess.