Use absolute paths in HTML files
Prepare for further improvements.  The main reason absolute paths are
wanted is for HTML file in other directory than /.  Without absolute
paths those file can't really access CSS and images.

We could have used custom relative paths but it will be a nightmare to
manage because we do not really control where each generated file are
going to be.  Plus it will be necessary when we'll want to add CGIs to
do page generation on-demand.

It now require an HTTP server to function.  A simple way of launching
one is using python builtin HTTP server:

	python -m http.server