Default database directory is now ".teerank" instead of "data"
"data" is usually used for static data in a website context.  By
renaming it to ".teerank" we also hide it, it may or may not be a good
thing in some situation but when database is generated inside the git
repo for developpement purpose it makes "ls" a little bit easier to read
since database and project file are not mixed up.