Do not update non-CTF servers as often as CTF servers
This patch aim to reduce the time taken by teerank-update, and more
precisely by teerank-update-servers.  With this patch, we are polling
100 servers in average, instead of 1000.  It greatly reduced the running
time of teerank-update.

However, it add some complexity to the system that may not be wanted.
So we will see in the long run if this pachset was a good idea.

It works by storing the last time the server was succesfully polled and
when the server need to be polled again.