Fix case when a server is seen for the first time
In this case we first try to read its state and return 0 if state is not found
After this commit if old server state was not found then 1 is returned
Motivation: When we see server for the first time there is not state recorded
yet. After read returns 1 then write occurs which writes the state. If read
returned 0 no write would ever occur.