Add a binary to remove offline servers
On teerank.com, more than 7000 servers has been piled up since the
launch of teerank.  Master servers can hold only 1024 servers, meaning
probably ~6000 servers are in fact offline.  So we don't need to track
them.

Commit bd065846 ("Do not update non-CTF servers as often as CTF
servers") aleady greatly reduced the number of servers polled every 5
minutes or so.  But offline servers will still be polled every 2 hours
or so.

This patch add a program to remove servers that have not been
succesfully polled since a given duration (in days).

A last note about this list of 7000 servers on teerank.com.  During my
test it seemed than far less than 6000 servers were in fact offline.  A
lot of them still answered my polls.  But I'm not sure of this.  We will
see in some weeks how the server's count is doing.