cgi: Files can have extra dependancies
Before only the source file (ie what was feeded into stdin), if any, was
considered as a dependency.  The issue is that source file doesn't
conatins all the dynamic information on a page, and therefor if one of
those infos change, the page will not be regenerated.

So now by adding a field to reference extra dependencies a file may
have, we are able to hopefully have a list of all files/directory where
informations on a page are stored in.

This should fix some issue about clans pages not fully up-to-date with
rank pages.