When verbose mode is on, print helpful messages
To have a better idea of what's going on when you run a program, this
patch introduce a verbose() function that works like printf() and only
print something when TEERANK_VERBOSE is set.

This messages are printed on stderr so it doesn't break piping.  They
are only there to make the whole projects easier to understand and
sometime provide some debug informations.