Move clans data in one file directly in $db/clans
Actually, to each clans a directory is associated.  In this directory, a
single "members" file contain a list of players name.  This patch
basically remove the directory and rename "<clan_name>/members" file to
"<clan_name>".  It's done for the same reasons exposed in commit
919a6dda ("Group every player's data into a single file").

Also this commit make teerank-update-clans use functions in player.h to
handle players.