Add a debug switch to help debugging
I'm usually not a fan of such addition but I haven't found *any* way to
retrieve the standard error output of fcgiwrap.  So this switch make the
cgi print error repoart in the response sent to the client.  That way
500 Inter errors came with the full error message.

Of course by default this switch is off, so that we don't show
potentially sensitive informations by default.