search: Do not display more than 50 results
Since we have no more than 50 results, we can just use an array of 50
results and remove/add results during the process.

Well, in fact we use a linked-list backed by a pool of 50 results so
adding/removing results are in O(1).  The code is a little bit complex
but we can now handle almost any queries.  So this commit enable search
for queries with 1 and 2 characters.

This commit also add code to sort results by their relevance.  If the
query match the beginning (or the end) of a player name, the it will be
at the top of the list.

We will add more rules later to have even better relevent results.