search: Sort results by rank
It seems to be a simple addition but in fact it is more complicated than
it seems.

We need to load the player, but we don't want to load every players.  So
we only load the player if he may belongs to the final results.

Also this commit fix some bugs as well, but I'm too put them in a
separate commit anyway.  One of them was that to_lowercase() was not
setting the terminating nul byte at the end of the destination string.