Update clans based on delat rather than full database
update-clans used to read every player entry and build a list of clans
and then write it to the disk.  Now we only change clans that have
changed and nothing more.

update-players now output a delta when a player have changed his clan.
This delta is fed to update-clans who then update the database.

The drawback of this method is when an error occurs, a player can be in
referenced by two clans, or even by no clans if update-clans failed.  So
we should probably introduce a new program to rebuild the list of clans
and their member from the full database, like the old update-clans.