Add a program to repair corrupted clans entries.
It just work like the old update-clans binary.  It look for all player
and build in memory a list of every clan with their members.  It then
update every clans in databse if data missmatch.

Note that it only update the clan when data missmatch, meaning no write
are performed if the entry on the disk is fine.  This is done to reduce
call to write() to the minimum, hence avoiding failure cases.