Move generate-*.c files in a separate directory
Since they are a special category of binaries that are supposed to be
used by CGI, it makes sens to put them in a separate directory.  Plus we
may have other binaries generating JSON instead of HTML, and in this
case they will be in their own src/json/ directory.