html-range-page: Take page number as parameter
Before the page file of the database was feeded on stdin.  But this
approach have some flaw the program may access other file on the
database, and use TEERANK_ROOT as the database root.  However the file
feeded on stdin could cam from another database.  So the result wouldn't be
consistent.

Instead everything is passed as paramter so TEERANK_ROOT is used no
matter what and we have a consistent behavior no matter what.

This policy is already in use on every other binaries in src/html