cgi: Remove cache
Tests showed that once database files are in OS file cache, every
request is handled within 0.001ms, cache or not.  So cache may speed
things up but it had some drawback we can't ignore.

First the code was overly complex because of the cache.  Questions like
where to put file in cache, how to know when cached file is outdated
make things really complex to deal with.  So by removing cache, we make
the code really simpler, and that's the main reason for this removal.

Also the cache messed up debugging because sometime cached files were
not marked as outdated and I was doing my test on the cached data,
without the CGI spawning any program.  By removing cache we open the way
to add automated testing.

Last but not the least, some configuration env vars were only used by
the cache: TEERANK_CACHE_ROOT and TEERANK_TMP_ROOT.  We can now remove
them.  Not a big deal but that's still nice.

Webservers often comes with ready-to-use cache, it may be even better to
use that (if necessary) instead of our old custom cache.