Add data structures wich doesn't store full historic
It aims to avoid calls to malloc() when we don't need to read the full
historic.  It also aim to to provide some sort of read-only structure
wich can only be read and not be written.  It makes the code simpler and
clearer by having one structure for each cases, and not one structure
for every use cases.