Merge format version 5 and 6
What this commit does is remove format version 6 and merge it's content
in format version 5.  What that means is that format version 5 is now a
completly new format than it was in former commits.

The reason for this merge is that managing formats is really cumbersome.
Each convertion program does needs to include old source for the old
format to be able to read it.  Hence to each new format some part of the
old source code should be keeped.

The previous and now deprecitaed old (implicit) policy about format was
that a new script has to be made each time format change, at whatever
point of the developpment.  That way any database created at any point
of time, even with developpment source code can be converted to the new
and most recent format.

The new policy still require a script to migrate old format to new ones,
however is does only enforce such script for *release* of teerank.
Implying now a correct versionning of the project should be done.

This new shiny policy does imply that database created with
developpement version of teerank (ie, non-release ones) cannot be
migrated to the new format once a release is made.  It means that only
releases should be used for production.

It also does means that a release is more or less a freeze of the
at-the-time format.  So some kind of support should made if some bugs
are found untile the next one.

This commit is *not* a release announcment, it just introduce the notion
of release for futur modifications.