historic: Add an upper limit for the number of records
Now user can specify the maximum number of records and historic can
hold.  If the historic is full when adding a records, the oldest record
is deleted.

Use this feature to limit hourly records to 24 and daily records to 30.

With this commit comes some changes on how historics are initialized,
and propagate this change to players.  I also had bugs and still some
binaries might not work as expected.  However changes are so huge that I
prefer to put them in a seperate commit for later.