elo-graph: Add a switch to generate rank graph too
With this change, "elo-graph" is not really the right name anymore.  It
should be changed in a later commit.

This change was not as straightforward as I thought, mainly because code
of elo_graph was really coded for graphing elo points.  Hence
introducing ranks which are an unsigned quantity required a lot of changes.