graph: Increase padding on the right to make things even
Horizontal axes are a little bit largen than the graph itself so thatif
the first or the last point overlap the axes, we can still see the axe.

However this extra size was touching the border of the picture.  This
commit add an extra padding to have an even padding on every side.