Make "ranks" file seekable, remove its pagination
Database of teerank.com have more than 700 pages, hence 700 files were
created - one for each page - from the main "ranks" file.  It was
necessary because ranks file was not seekable and so to found a
particular player we would have to go through every players.

To avoid this huge traversal each time a page had to be displayed,
"ranks" file was splitted in pages.  Hence showing a page was just a
matter of opening the right file and dump its content.

As said above, with 100 players per page, 700 page files are created for
a 70000 players database.  It is already a lot of files, plus a huge
part of them were never used.  For better performance and maybe more
simplicity, it had to change.

This commit solve the issue by making "ranks" file seekable.  The change
is pretty straightforward, as it is just a matter of padding player's
name so they have a fixed size.  printf() handle that padding for us, so
a huge part of this commit is in fact writting the upgrade code to
upgrade an old database to the new format.

Since page files are now useless, this commit also remove the pagination
mechanism.