Merge "state" file and "meta" file
Servers required 3 files: server directory, meta file and state file.
By merging state file and meta file, not only server data can be stored
in a single file, but it make the code simpler because code do not have
to make sure meta and state data are synched.

The change is everything but trivial, because add-new-server,
remove-offline-servers and update-servers used struct meta heavily, or
relied on different assumptions of meta and state files.