Merge elo historic and rank historics
The rational for this change is that elo and rank are highly coupled.
When plotting elo and rank, we want them to be displayed together so we
can correlate elo changes to rank changes.

However, since rank can change is the player do not player, the last
rank is always stored outside the historic so we can keep track of it.

Since elo score is compited first and then rank is computed, when elo
score is inserted in historic, the rank is set to unranked.  When a new
rank is set, if the last record of the historic is unranked, the new rank
become the rank associated with the preciously unranked record.  Hence
set_elo() and set_rank() are little bit different.

New historic does not have an upper limit for now, because some
testing showed that only a couple of players do have a lot of activity,
so database will not be filled very fast.

Also this change does *not* show elo and rank on the same graphic.  This
change will likely comes in a further commit.