Generate script-header.inc.h
the idea is to have only one place were default config value are sets.
This place is now default_config.h, and thanks to xmacros, c code is
generated and C code that generate bash code is generated too.

That way a change to default_config.h does update C programs as well as
shell scripts.