Merge player_exist() and read_player()
player_exist() was not really satisfying because often it is followed by
read_player().  So it is better to just check fopen() return value in
read_player() to know if a player exist or not.

The code seems to be better, also it kind of break with actual
conventions: read_player() will return PLAYER_FOUND, or 0 on success,
but every other function in player.h and in teerank in general does use
0 for failures.  I don't really mind tho because it is still better than
player_exist() and the even older implicit player creation on failure.