historic: Remove unused HISTORIC_ZERO, update documentation
In particularily, documentation of init_historic() was outdated as it
stated that init_historic() has to be called for every new or re-used
historics.  Which is not the case anymore: init_historic has to be
called once for all on an historic.

Last but not the least, this commit also add missing documentation of
create_historic().