cgi: Do not fork to generate pages
No functional changes, exceptcgi is now one big binary capable of
generating every pages.  Before getting into details, a little bit of
historic.

For simplicity and better testing, I used to make every pages a
standalone program.  Hence there was a program for generating rank pages,
one for clan pages, and so on.  When I did create CGI, I just used a
fork() and execl() and did some dup2() to generate the page.

This worked well, with one drawback tho: we needed to make sure binaries
was reachable from $PATH.  Performance would have been better without
any fork(), but given the little audience of teerank, it was a
non-issue.

So now, no more standalone binaries for each page.  It means that
instead of doing a fork(), we just call a function for the wanted page.
Of course we still need to do some pipe(), dup() and dup2() to handle
error as nicely as possible.  Although it may not be needed if pages
were written carefuly, I'm not sure.

Standalone binaries was still handful for debugging, so I might
re-introduce them in a later commit.

Let's note that load_config() does not need to be called at the
beginning of each page function, because it is done at the start of the
CGI.

Routing stuff has been a little bit splified using a macro.  "file"
structure has been renamed to "page".

I'v done some basic benchmark to see how fork() did affect
performances.  I generated a page for each player in database (107 at
the time).  It took 0.650ms before the commit, and 0.500ms after.  Not
that great, but still a nice "side-effect".