upgrade-4-to-5: Set historic nrecords to zero
Not doing so was triggering an assertion saying that nrecords > length,
because nrecords was not reset between two players.

What's strange tho is that it never showed up in my testing, I had to
attempt an upgrade of teerank.com database to catch this one.  It is
strange.

Good thing is that an assertion catched this bug.  I wonder if I should
had more assertions to teerank.  By that I mean more detailed
assertions, not just checking if pointers are NULL.  We will see.