skip_record: Fix a warning about unused return value
Actually this warning poped when I tried to build teerank for
test.teerank.com, to do some testing.  GCC version are the same, so it
is a mystery why I don't have this warning of my computer.  Plus I
already compiled this code on the testing machine many times before and
never got the warning...

But oh well, this should fix it.