Make a upgrade before upgrading
The rational behind this change is that when an upgrade fail, database
is left in an unknown state.  It means that re-doing an upgrade on it
will probably not work.  So if the user did not make a backup before
upgrading, he or she is screwed.

Instead of asking the user to make a backup and waiting for his
confirmation, we just take the initiative to make a fast backup and then
restore the database on failure.

Backup is cleaned at exit to keep the directory clean.

This commit does also remove the redicrection of echo to standard
error.  before teerank-upgrade was not designed as an interactive tool,
but now that it require user input in some situation, it is an
interactive tool.  Hence message should be send to standard output
because they are part of the interface.