Add a sitemap.xml file for search engines
I don't know sitemap.xml file are used by search engine, and I don't
want to create an account for teerank for each search engine?  So we
will if this file have any use at all.

What I hope is that search engines use it to also put links to CTF and
About on search results.