Add a function to handle database paths
Using snprintf() each time we wanted to build a path was redundant,
annoying and error prone.  It turns out almost every snprintf() in the
codebase are for database path.  This commit add a little function to
handle path creation and every possible errors.

teerank-upgrade-4-to-5 has not been updated because I'm not sure if we
should merge a change like that for upgrade programs...