Use static library of previous stable release upgrade binary
(Bump version from 2 to 3)

The idea is to only maintain the upgrade binary to upgrade from the
previous stable version to the current version.  Also, the idea is to
remove duplicate code, like code for reading players, ...

To achieve that, we use static library built from stable versions, and
prefix every symbols with "teerank$(VERSION)_".  Hence upgrade programs
will use teerank5_read_player() followed by some convertions and then a
write_player() (or teerank6_write_player(), but in that case they will
have to declare it as well)

One things tho that is not yet really nice is that such functions will
have to be delcared because header files are not importedd and converted
to match the prefixed symbols name.

But that's still better to copy paste code for each update programs.
For instance, an upgrde program that want to upgrade players will look
something like this:

	/* Header files are not imported, this structs and functions needs to
	 * be declared */
	struct teerank5_player { ... };
	int teerank5_read_player(
		struct teerank5_player *player, const char *name);

	int main(void)
	{
		struct teerank5_player src;
		struct player dst;

		teerank5_read_player(&src, "00");
		upgrade_player(&src, &dst);
		write_player(&dst);
	}

So now upgrade/ directory will only contain the upgrade program to
upgrade a database from the previous version from the current one.  The
content of this directory must be wiped at each version bump.

To conclude: this commit allow more code to be shared accross release
and reduce maintenance cost for upgarde programs.