Automatically Prefix headers of the previous release
To use static library of the previous teerank release, we needed to
import its header files without any name conflict with header files of
the current release.

On way I actually tested was to prefix every global symbol by hand,
using some preprocessor trick, like:

	int foo(const char *bar);

would become:

	int __(foo)(const char *bar);

where the macro __(name) would prefix the given name when necessary.
But this approach make header files almost unreadable, and it was not
really working anyway.  For instance how can you prefix macro when they
are defined?

Eventually I came up with the idea of having a program that
automatically prefix header files for me.  This commit is the
implementation of the idea.  Since I was not going to write a full C
parser, I just wrote some sort of simplified parser.  It doesn't handle
the full C syntax but it should be enough for header files.

It has the advantage to keep file structure and comments.  It has the
disavantage to rely on assumptions that may not be true for futur
release.

The good news is that upgrade programs will be much simpler to write and
maintain.  Now they pretty much just have to write a function to convert
old data structure to the new ones, and every IO stuff is nicely
handled.