Store the last seen date of a player
Actually the date is not update when the player is seen.  It is only set
at player creation.

This change is a good opportunity to test the new upgrade
infrastructure.  Since player data format changed, it was necessary to
write an upgrade program to upgrade old data to the new format.  It is
now surprisingly easy, it basically consist to write the convertion
function from old data structure to the new one.