Write index in JSON
Required some modification to JSON code so that array can be seeked.
For that purpose some code has been added to make sure every entry of an
array have the same size, so we can compute the seeking offset.

We needed to store entry size in scope, as well as a way to compute
scope size when closing it.  The latter is achieved by storing current
file position when starting a new scope, and substracting it to the file
position at the and of the scope.

Also wee needed a way to compute at compile time the maximum size of a
json field.  This is what macros JSON_*_SIZE does, replacing the old set
of macros we had previously.

Overall JSON code became quite complex.  Espcially error lanagement and
corner case make it a little bit tricky to follow.  I want to make it
simpler before marging it into the master branch.