json: Allow grouping calls without error checking
The idea is to make the json API easier to use.  Checking for errors
when necessary for each function call, as shown in the following
example:

	if (!json_write_int(&jfile, NULL, foo))
		return 0;
	if (!json_write_string(&jfile, NULL, bar, sizeof(bar)))
		return 0;

Using a condition each time was cumbersome and very redundant.  This
commit make the above code easier to write, by only checking for erros
at the end:

	json_write_int(&jfile, NULL, foo);
	json_write_string(&jfile, NULL, bar, sizeof(bar));

	if (json_have_error(&jfile))
		return 0;

Alhtough the latter example is larger the former one in this particular
case, the overall code is much easier to read.  It is worth noting that
it is still possible to write code like the first example, but it's not
recommended.

This change is not trivial as it required to make the API resistant to
failure case detected in previous call.

Overall 100 lines are removed by this commit, so I would say this is a
big win.