Document the JSON API
Link bottom tabs to this documentation.  And remove dummy .json.html
pages.

The plan was to have one documentation page for one JSON page, but on a
second though I think it was better to put all the documentation on the
same page, because it makes the whole API discoverable.