Implement all JSON pages
It feels a bit dirty as it is because we are not doing proper parsing to
be as fast as possible.  For example we only check for ',', ']' and '}'
to count fields, and we only check for strings when we want to remove
spaces between fields.

It does works but that is also confusing for sure:

	dump_n_fields(file, 4);
	putchar('}');

The above code does only print the first 4 fields of a player, to avoid
printing the whole historic.  We need to close the object manually
because dump_n_fields() stop as soon as it counted up to 4.  Fast but ugly.