Handle HTTP 301 error code
HTTP 301 is "Moved Permanently".  It is when a oage changed its
location.  For that error code to function, we need to add one extra
header field with the new url ("Location: <new_url>").  It is not a
classic erro code and therefor cannot be used with error().  We needed
to add a function redirect() for this purpose.