Prevent overflow when converting IP to string
Technically I'm sure it is not possible to overflow but this commit add
safeguards using assertions.  You may wonders if this is really
appropriate to use assertion on user supplied input but in this case IP
and port are provided in a fixed size buffer.  So this commit only
really make sure our code is right, not user supplied data.