Only show CTF servers
In a bad attempt to reduce the dusplay of unregistered servers.
Obvisouly the right fix is to register every player no matter the
gametype they are playing on.  But in any case, one expect only CTF
servers in the CTF section.

There is still unregistred players because unrankable players are not
always put in the database, I need to look into that one as well.