server: Set last seen to NEVER_SEEN on creation
This commit also set server->expire to zero.

The rational behind this change is to have a consistent lastseen value
when the server is created.  Before lastseen was set to the actual
time(), but it wasn't accurate: on creation the server has never been
contacted so we can't really assume it is online.