cgi: Remove unused page main function
I added those pages while experimentaing with JSON API.  I though it
would have been a good idea to have a dedicated page of JSON
documentation for each query type.  However after testing it I disliked
it and removed it.  Since there is not a lot of JSON query available (7)
they all fit in the same page of documentation, and by looking at the
page summary we can have a nice overview of the whole API.

So I removed it but forgot about this declarations.