index: write() and read() data instead of using JSON
Parsing JSON was actually taking a lot of time.  On a regular search in
a database of 87K players, 0.13s on 0.25s are saved just by this change.
It may even be better to load the whole index all at once instead of
read()-ing every entry.

It also makes the code really simpler.