index-page: Use mmap() instead of read()
It seems to be faster on my machine with a traditional HDD.  It may be a
different story on SSD.

I actually tried to malloc() and read() the whole index page, but it was
slower: 0.040s vs 0.022s with a static build with musl.  However both
are faster than the old code, that is read()-ing each entries.