Add a status page
This page show the status of the whole teerank website as well as the
status of the JSON API.  Even tho for now none of the displayed status
are correct, they will be linked to there real state in later commit.

This page also show the status of teeworlds masters.  Again, the status
actually displayed are just placeholders.

This new page cna be accessed from the footer or by clicking on the
alert banner introduced in commit 36d54ae2 ("Show a warning when
database updates stopped").