Use pool for polling masters
Pools have to be modified to make it work for masters servers polling.
First, for one packet sent, masters can answer with one or more packets,
so we need a way in the pool to keep poling entries.  Also, on failure,
we don't want to sent a new pakcet again to avoid duplicates.

It seems to work pretty much okay.