Allow index page for root route
strtok() does not yield empty string, so "/players/" and "/players" are
handled the same.  However, we would like to have a page for "/" or "",
but strtok() will immediately return NULL.  So if the first call to
strtok() return NULL, we considere the return value to be an empty
string instead.