Fix potential infinite loop when polling servers
After some time running teerank-update, teerank-update-server got stuck
in an infinite loop.  I'm still not sure why this commit fix it, but the
logic was wrong anyway.

By default, a pool entry that just received data remains in the pending
list, until the user explicitely remove it.  This behavior is useful
when polling master servers, because they send multiple packets as an
answer.

However, the actually removed the server from the pending list, setting
it back to IDLE.  I'm sure in nasty circumstences there is a way to
infinitely add and remove servers to the pending list, resulting in some
sort of infinite loop.

Fix it by just removing the bogus remove_pending_entry() after it
received an anwser.