html_footer(): Fix mis-use of html() causing infinite loop
This one is actually silly: html() expect a starting tag without any
space before.  But actually I did add one space before and it resulted
in mis-indentation underflowing when closing <html>, causing an (almost)
infinite loop.

I didn't catch it earlier because I was building in release mode, and
indentation is disabled in this mode...