Use a custom function to calculate elapsed time
The way it was done before this commit could yield strange results and
was potentially slow.  It basically computed the difference in second
before converting it again in struct tm.

	sec_elapsed = mktime(b) - mktime(a);
	elapsed = *gmtime(&sec_elapsed);

"elapsed" contains here a date computed from UNIX epoch, and it was used
at the broken down elepased time.  It kinda works but gmtime() may be
slow, and output is not really regular, as month length will not be the
same.

The "fix" is to just assume months have all the same length when we
speak of elapsed time.  By not using gmtime() we make sure our output
come in constant time rather reliying on a potentially complex gmtime().

So our function starts by defining length of every unit of time:

	const time_t MINUTE = 60;
	const time_t HOUR = 60 * MINUTE;
	const time_t DAY = 24 * HOUR;
	const time_t MONTH = 30.4 * DAY;
	const time_t YEAR = 12 * MONTH;

And then use it to broke down the given elapsed time (see the code for
the full algorithm).