Rework pool.c to make it simpler
This pece of code is complex and it was buggy: in some unknown
circumstances, it could go in an infinite loop.  This rework aim to
remove any superfluous codepath.

Before, an entry could belongs to two linked list: the main list and the
pending list.  It contained a variable to indicate the state of the
entry: IDLE, PENDING, POLLED, FAILED.

While it was already quite simple, it turned out that we could make it
even simpler:  Each entry now belongs to one and only one of those
lists: idle, pending, failed.  While getting ride of the status
varaiable, we also get ride of the code handling two different lists.

By doing that we also get ride of the function to iterate over idle
entries, because before we had to filter out polled, failed and pending
entries from the main list, while now we just have to iterate through
the idle list.  Same for failed entries.

We also handle correctly entries with answer spread over multiple
packets (like master servers for instance).  When an entry got an
answer, it is marked as "polled" and will stay in the pending list until
it is manually removed or until the poll failed.  In the latter case,
since the entry still received data despite the poll failure, it is
considered as a success.

This is necessary because we cannot reliably know when a master server
will stop sending new packets.