Moving to SQlite 3
Using a custom database seemed at first glance a good idea, but now that
I've done the SQlite port, it appears as a very bad idea in retrospect.

On the good side, it was really fun to program, and I designed it in a
way that displaying a page in the cgi was pretty much just like reading
single file.  It was really efficient for reading.

That said, this commit shows that it tooks a lot of code to get it
working.  it was a pain in the ass to maintain such database.  First
because it is not crash-proof.  I had -even in the latest version- a
lots of file corruption, for unknown reasons.

Also, having 150k+ files does takes a lot of disk space.  While the
linux kernel handles that without any issues, it became clear that it
was going to be a big problem if I wanted to scale up to 1M files.

Having a relational database will make it easy to filter and query data
to fit our needs, without much efforts.  Where a single change in the
logic before required to changes all the underlying data structures.

The code is now vastly simpler, and the database should be more
consistent and safe to unexpected crashes.  That said, displaying a page
is not that efficient as it was before.  Up to 0.040s on my computer to
display a clan for a 9k players database.  Where it was taking 0.004s
with the old one.

In fact, the whole point of this project was to have a real, big enough
dataset to compare raw database and SQL ones.  Mission accomplished, and
in that case, SQL clearly won the battle.

For now I'm still concerned of performance issues, I will try to
optimize everything I can regarding sqlite to get back to decent
rendering time.

On a side note: I sometime have "database is locked" errors when
updating, and I'm not sure why.  I've added a 5s timeout so it is very
unlikely a race between update-servers and update-players, even tho it
looks like it.  I plan to merge update-servers and update-players in a
single process so hopefully it will fix that issue along the way.

Anyway, that was the biggest commit I've ever done for sure.  And the
good news is that it actually *removes* a lot of lines, while preserving
100% backward compatibility.  Woohoo!