Introduce a nicer way to loop over masters
The SQlite API can be quite verbose and repetitive, especially for
select request.  This commit aim to abstract SQlite calls in a nice and
readable way:

	if (sqlite3_prepare_v2() != SQLITE_OK)
		goto fail;
	while ((ret = sqlite3_step(res)) == SQLITE_ROW)
		...
	if (ret != SQLITE_DONE)
		goto fail;
	slite3_finalize(res);
	return;
fail:
	/* Error handling */
	sqlite3_finalize(res);

Now becomes more concise and explicit:

	foreach_master(query, &m)
		...

	if (!res)
		/* Error handling */

Even tho the foreach() macro is mainly preprocessor hackery, I think it
is worth it and remove a lot of boilerplate code.  The logic of this
foreach has been abstracted and put in a new file: 'database.h'.

Also introduce extended master, wich are like master, but with
"nservers" set.  Before you would write:

	"SELECT" ALL_MASTER_COLUMN "," NSERVERS_COLUMN

While now you will prefer:

	"SELECT" ALL_EXTENDED_MASTER_COLUMN

Coupled with foreach_extended_master(), it is more consistent than
having special case for "nservers" for every function call.

Players, clans and servers are not yet migrated to this new API because
its not as trivial as for masters.  First, we will need to get ride of
read_player() and read_server(), becuase they are easily replaced by
foreach() API.  Then we will need to accomodate for query that use
bindings.