Reuse last query in exec() and foreach_row()
Some piece of code executed the same query multiple times with different
bindings.  To avoid re-preparing the query each time, it was necessary
to keep track of the previous query and reset it instead of re-preparing
it.

What's cool is that the function prototype do not change at all, making
this change transparent to the user.

Also, this change remove the need for a custom break_foreach, because
query finalizing is delayed up until the next call to foreach_row().