Compute ranks only once every 5 minutes
We can't calculate ranks at every update/insert, this is just too
expensive.  For instance it tooks more than 2 seconds to update ranks of
50 000 players.

So just do it every 5 minutes or so.  This commit is bigger than one
might expect because we also have to take care of a bunch of other
things to deliver consistent data to the end user.

Indeed, we need to delay elo score update just before ranks are
computed.  Not doing it can result in many inconsistencies.  For
instance a player could get a better rank with less elo point than
another player, just because his rank has not been updated to reflect
his points loss.