Drop rank indices while re-calculating ranks
We don't want the indices to update at every UPDATE queries, so disable
them to speeds things up.  I tested it on a database with 10 000 players
and I  saved up to 20% of the running time of recompute_ranks().

Also add an index with the same order than the one used to rank
players.  It also speeds things up quite a lot.