Show a message when the database can be opened readonly
If the database cannot be opened, we try to create it, hence there are
no reasons to show an error message.  If and only if the creation failed,
then an error message is printed.

However, when we want to open a database in read-only mode, if opening
failed, we don't try to create the database.  So display the error
message right away.