Merge *-list.json.c into "list.json.c"
It will ease maintenance of the JSON API, because it's in a deplorable
state.  There are still a lot of code shared between the HTML list and
the JSON list.

Ah, and this commit just drop the retro-compatibility of the JSON API.
It wasn't working anyway so I hope nobody care.  The URL now match the
HTML counter part.