Merge page setup callback and main callback
What's was done in steup callback is now done in main callback.  There
was no reason anymore to keep them splited.  Plus, NULL values couldn't
be passed to main callback because args had to be strings.

This get ride of a lot of code who was parsing argc/argv.  This commit
also makes the routing code much simpler.

This comit introduce one caveat: redirect() was not supposed to be
called in main callback and it's not working anymore.  I did not fixed
it yet because I may rework how main callback can set HTTP header
fields.  But for now this failure only affect old URLs.